+++ src/components/SwapLayout/Header/index.tsx
@@ -7,12 +7,17 @@ import { formatSeconds, dateToBatchId } from 'utils'
 import UserWallet from 'components/UserWallet'
 
 // Header: Subcomponents
-import { NavigationLinks } from 'components/Layout/Header/Navigation'
-import { HeaderWrapper, BatchNumberWrapper, CountDownStyled, TopWrapper } from 'components/Layout/Header/Header.styled'
+import { NavigationLinks } from 'components/SwapLayout/Header/Navigation'
+import {
+  HeaderWrapper,
+  BatchNumberWrapper,
+  CountDownStyled,
+  TopWrapper,
+} from 'components/SwapLayout/Header/Header.styled'
 
 // hooks
-import useNavigation from 'components/Layout/Header/useNavigation'
-import useOpenCloseNav from 'components/Layout/Header/useOpenCloseNav'
+import useNavigation from 'components/SwapLayout/Header/useNavigation'
+import useOpenCloseNav from 'components/SwapLayout/Header/useOpenCloseNav'
 import { useTimeRemainingInBatch } from 'hooks/useTimeRemainingInBatch'
 import { HelpTooltipContainer, HelpTooltip } from 'components/Tooltip'
