+++ src/pages/About.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { ContentPage } from 'components/Layout/PageWrapper'
+import { ContentPage } from 'components/SwapLayout/PageWrapper'
 import addressesByNetwork from 'api/deposit/batchExchangeAddresses'
 import { getNetworkFromId } from '@gnosis.pm/dex-js'
 import { EtherscanLink } from 'components/common/EtherscanLink'
