+++ src/pages/Trade.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import TradeWidget from 'components/TradeWidget'
-import { PageWrapper } from 'components/Layout/PageWrapper'
+import { PageWrapper } from 'components/SwapLayout/PageWrapper'
 
 const Trade: React.FC = () => (
   <PageWrapper>
